rasp/requirements.txt
pyserial
-e git+https://github.com/niolabs/python-xbee.git#egg=xbee
paho-mqtt
